fix: migrations, clean-up
