+++ ayon_server/helpers/deploy_project.py
@@ -5,7 +5,9 @@ from typing import Any
 from ayon_server.auth.session import Session
 from ayon_server.entities import ProjectEntity, UserEntity
 from ayon_server.entities.models.submodels import LinkTypeModel
+from ayon_server.entities.project_skeleton import ProjectSkeletonEntity
 from ayon_server.events import EventStream
+from ayon_server.exceptions import NotImplementedException
 from ayon_server.lib.postgres import Postgres
 from ayon_server.logging import logger
 from ayon_server.settings.anatomy import Anatomy
@@ -193,4 +195,50 @@ async def create_project_from_anatomy(
         project=project.name,
         user=user_name,
         description=f"Created project {project.name}",
+    )
+
+
+#
+# Skeleton projects
+#
+
+
+async def create_project_skeleton_from_anatomy(
+    name: str,
+    code: str,
+    anatomy: Anatomy,
+    *,
+    library: bool = False,
+    user_name: str | None = None,
+    data: dict[str, Any] | None = None,
+    assign_users: bool = True,
+) -> None:
+    project_data = anatomy_to_project_data(anatomy)
+
+    project = ProjectSkeletonEntity(
+        payload={
+            "name": name,
+            "code": code,
+            "library": library,
+            **project_data,
+        },
+    )
+
+    await project.save()
+
+    await EventStream.dispatch(
+        "entity.project_skeleton.created",
+        sender="ayon",
+        project=project.name,
+        user=user_name,
+        description=f"Created project {project.name}",
+    )
+
+
+async def promote_project_from_skeleton(
+    project_name: str,
+    anatomy: Anatomy | None = None,
+) -> None:
+    raise NotImplementedException(
+        "Promoting project from skeleton is not implemented yet"
     )
